backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# API Clients & HTTP
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# OpenAI (for LLM-based CAS parsing)
openai==1.58.1

# Scheduling
apscheduler==3.10.4

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# PDF Processing (for CAS parsing)
PyPDF2==3.0.1
pdfplumber==0.10.3
camelot-py==0.11.0

# Date/Time
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# Financial Calculations
pyxirr==0.10.4

# Logging
loguru==0.7.2

# CORS
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Crypto (for API signing)
cryptography==41.0.7
